calculateRevisions was unable to run, to fix, I moved the function into a module of its own and did the necessary imports in other files
